chore: rename folder from PascalCase to lowercase
